Move all project header files from 'include' to 'src' directory
The header files (.h) for the project used to reside in the 'include'
directory that was at the root of the project. These files have now been
moved to their respective folders inside the 'src' directory (root/src). This is
mainly to reduce maintennce between the header files and cpp files as
well as making the project structure a bit more cleaner.