Add new CSS tokenizer
This new tokenizer for css replaces the old CSSParser(src/Parser) class as how this
web engine will tokenize and eventually parse the CSS. The old CSSParser
was very ad-hoc.

This new tokenizer has been written to the CSS3 spec.
Reference: https://www.w3.org/TR/css-syntax-3

There are a few tokens not supported, but will be implemented as the CSS
tokenizer and parser matures.