Start working on rendering pipeline and refactor browser window
This change introduces a new rendering pipeline for the render tree, it
performs a compute phase to calculate all CSS box dimensions then
renders the box as well as any content. In the future, rendering the
content will be a seperate phase.

The other change is making the browser window seperate as we need space
to add GUI elements such as menu bars.