CSS: Add entire StyleRules to an Element
This is much needed bug fix as selector specifity was not working as
defined. Instead of an Element holding one ComplexSelector, it now
stores a list of StyleRules. These style rules are looped over in order
to override overlapping declarations.

A single StyleRule now contains A, B and C values that are additions of
all A, B and C values in a single ComplexSelector from the list.