Move HTML Lexer's helper functions into own class
We have now moved some of the helper functions that classes were
previously referencing the Lexer class for to a new purpose built class.
This new class is called LexerUtils and sits at the root of the project.
Inside this class we have a few functions that we'd like to use across
the project so it's better that they're in their own namespace rather
than relying on the Lexer.

Worth noting that the Lexer is only for HTML
parsing and shouldn't be used a tool kit.

We have also removed the keyword "const" from shared_ptr<> parameters.
This is in preparation for when we re-write how shared pointers are used
across the entire project (fun, fun, fun).