Smart pointer refactor
This is quite a substantial commit with a few key changes. The main one
is that we've changed how smart pointers are used in the HTML
tokenizer\parser as well as the CSS tokenizer\parser. This includes
ensuring that a child node does not own it's parent so instead of
storing the parent in a shared ptr, we now store it in a weak pointer.
This applies to DOM nodes and LayoutBox nodes.

Another key change is that we now store the Document in the Blitz
instance once a certain stage of the engine as finished modifying it.
This is then passed to the next stage. e.g. HTML Tokenizer > HTML
Document Parser > Selector Engine > Style Resolver > WebView

That's the bulk of, other changes a are just minor style changes and
clean up.